Refactor: create a with state hoc
