+++ src/components/bookmark-button/index.tsx
@@ -2,24 +2,18 @@ import { IconBookmarkAdd16 } from '@hashicorp/flight-icons/svg-react/bookmark-ad
 import { IconBookmarkRemove16 } from '@hashicorp/flight-icons/svg-react/bookmark-remove-16'
 import { AUTH_ENABLED } from 'hooks/use-authentication'
 import Button from 'components/button'
-import { withDialog } from './sign-in-dialog/with-dialog'
+import { withState } from './with-state'
 import { RemoveBookmarkIcon, AddBookmarkIcon } from './icons'
 import s from './bookmark-button.module.css'
 
-/**
- * TODO
- * - create a 'plugged' in HOC that can render modals, update bookmark state etc.
- * This HOC will be used by this button and the one in tutorial meta
- */
-
 export interface BookmarkButtonProps {
 	isBookmarked: boolean
-	openDialog(): void
+	handleOnClick(): void
 }
 
 function BookmarkButtonIconOnly({
 	isBookmarked,
-	openDialog,
+	handleOnClick,
 }: BookmarkButtonProps) {
 	// NOTE! - hiding this component from prod until auth is enabled
 	if (!AUTH_ENABLED) {
@@ -29,12 +23,7 @@ function BookmarkButtonIconOnly({
 	return (
 		<button
 			aria-pressed={isBookmarked}
-			onClick={() => {
-				// TODO use the create / destroy methods in the client
-				// or render dialog to prompt auth if not auth'd
-				// if not auth'd, show dialog
-				openDialog()
-			}}
+			onClick={handleOnClick}
 			aria-label={helpText}
 			className={s.button}
 		>
@@ -43,9 +32,9 @@ function BookmarkButtonIconOnly({
 	)
 }
 
-function BookmarkButtonWithText({
+function BookmarkButtonTextAndIcon({
 	isBookmarked,
-	openDialog,
+	handleOnClick,
 }: BookmarkButtonProps) {
 	// NOTE! - hiding this component from prod until auth is enabled
 	if (!AUTH_ENABLED) {
@@ -56,12 +45,10 @@ function BookmarkButtonWithText({
 			color="secondary"
 			text={isBookmarked ? 'Remove Bookmark' : 'Add Bookmark'}
 			icon={isBookmarked ? <IconBookmarkRemove16 /> : <IconBookmarkAdd16 />}
-			onClick={() => {
-				openDialog()
-			}} // TODO hook up to real state mgmt, show dialog etc.
+			onClick={handleOnClick}
 		/>
 	)
 }
 
-export const TutorialCardBookmarkButton = withDialog(BookmarkButtonIconOnly)
-export const TutorialMetaBookmarkButton = withDialog(BookmarkButtonWithText)
+export const TutorialCardBookmarkButton = withState(BookmarkButtonIconOnly)
+export const TutorialMetaBookmarkButton = withState(BookmarkButtonTextAndIcon)
